Added TransactionPage and HomePage
Transactions Page is same as previous Portfolio Page. Portfolio Page will show something else in future
