+++ src/App.js
@@ -4,6 +4,7 @@ import MainPage from "./Pages/MainPage";
 import LoginPage from "./Pages/LoginPage";
 import ResponsiveAppBar from "./components/ResponsiveAppBar/ResponsiveAppBar";
 import DataFetch from "./components/DataFetch";
+import HomePage from "./Pages/HomePage";
 
 
 function App() {
