+++ src/Pages/HomePage.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+function HomePage() {
+  return (
+    <>
+      <h1>HomePage</h1>
+      <h2>WatchList</h2>
+    </>
+  );
+}
+
+export default HomePage;
